+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Fixes
 
-* Fixed incorrect message aging in `GuildMessageChannelBehavior#bulkDelete`
+* Fixed GuildMessageChannelBehavior#bulkdelete  incorrectly deleting messages older than 14 days
 
 
 # 0.6.2